misc: format files with uncrustify 0.78.1

+++ modules/ui/qt/status.cpp
@@ -36,7 +36,7 @@
 namespace sight::module::ui::qt
 {
 
-enum class status_color : std::size_t
+enum class status_color : std::uint8_t
 {
     green,
     orange,
@@ -105,7 +105,7 @@ void status::configuring(const config_t& _config)
         const auto label_status_config = config_labels.get().equal_range("name");
 
         // Fill the label_status vector
-        for(int i = 0 ; const service::config_t::value_type& v : boost::make_iterator_range(label_status_config))
+        for(int i = 0 ; const service::config_t::value_type & v : boost::make_iterator_range(label_status_config))
         {
             const auto label = v.second.get<std::string>("");
             auto* status     = new QToolButton();
@@ -232,7 +232,7 @@ void status::change_nth_to_green(const int _index)
         _index < 0 || _index >= int(m_label_status.size())
     );
 
-    if(_index >= 0 && _index < int(m_label_status.size()))
+    if(_index >= 0 && _index < m_label_status.size())
     {
         const auto status = m_label_status.at(_index);
         status->setIcon(icon(status_color::green));
@@ -249,7 +249,7 @@ void status::change_nth_to_red(const int _index)
         _index < 0 || _index >= int(m_label_status.size())
     );
 
-    if(_index >= 0 && _index < int(m_label_status.size()))
+    if(_index >= 0 && _index < m_label_status.size())
     {
         const auto status = m_label_status.at(_index);
         status->setIcon(icon(status_color::red));
@@ -266,7 +266,7 @@ void status::change_nth_to_orange(const int _index)
         _index < 0 || _index >= int(m_label_status.size())
     );
 
-    if(_index >= 0 && _index < int(m_label_status.size()))
+    if(_index >= 0 && _index < m_label_status.size())
     {
         const auto status = m_label_status.at(_index);
         status->setIcon(icon(status_color::orange));
@@ -283,7 +283,7 @@ void status::toggle_nth_green_red(const int _index, const bool _green)
         _index < 0 || _index >= int(m_label_status.size())
     );
 
-    if(_index >= 0 && _index < int(m_label_status.size()))
+    if(_index >= 0 && _index < m_label_status.size())
     {
         const auto status = m_label_status.at(_index);
         status->setIcon(_green ? icon(status_color::green) : icon(status_color::red));
